Pipe: Refactor pipe related packages
