+++ iotdb-core/confignode/src/main/java/org/apache/iotdb/confignode/manager/pipe/task/PipeConfigNodeSubtaskExecutor.java
@@ -17,10 +17,10 @@
  * under the License.
  */
 
-package org.apache.iotdb.confignode.manager.pipe.execution;
+package org.apache.iotdb.confignode.manager.pipe.task;
 
 import org.apache.iotdb.commons.concurrent.ThreadName;
-import org.apache.iotdb.commons.pipe.execution.executor.PipeSubtaskExecutor;
+import org.apache.iotdb.commons.pipe.task.execution.executor.PipeSubtaskExecutor;
 import org.apache.iotdb.commons.utils.TestOnly;
 
 public class PipeConfigNodeSubtaskExecutor extends PipeSubtaskExecutor {
