+++ iotdb-core/confignode/src/main/java/org/apache/iotdb/confignode/manager/pipe/task/PipeConfigNodeTaskStage.java
@@ -21,8 +21,6 @@ package org.apache.iotdb.confignode.manager.pipe.task;
 
 import org.apache.iotdb.commons.pipe.task.meta.PipeTaskMeta;
 import org.apache.iotdb.commons.pipe.task.stage.PipeTaskStage;
-import org.apache.iotdb.confignode.manager.pipe.execution.PipeConfigNodeSubtask;
-import org.apache.iotdb.confignode.manager.pipe.execution.PipeConfigNodeSubtaskExecutor;
 import org.apache.iotdb.pipe.api.exception.PipeException;
 
 import java.util.Map;
