+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/task/builder/PipeDataNodeTaskBuilder.java
@@ -26,11 +26,11 @@ import org.apache.iotdb.commons.pipe.config.constant.SystemConstant;
 import org.apache.iotdb.commons.pipe.task.meta.PipeStaticMeta;
 import org.apache.iotdb.commons.pipe.task.meta.PipeTaskMeta;
 import org.apache.iotdb.commons.pipe.task.meta.PipeType;
-import org.apache.iotdb.db.pipe.execution.PipeConnectorSubtaskExecutor;
-import org.apache.iotdb.db.pipe.execution.PipeProcessorSubtaskExecutor;
-import org.apache.iotdb.db.pipe.execution.PipeSubtaskExecutorManager;
 import org.apache.iotdb.db.pipe.extractor.dataregion.DataRegionListeningFilter;
 import org.apache.iotdb.db.pipe.task.PipeDataNodeTask;
+import org.apache.iotdb.db.pipe.task.execution.PipeConnectorSubtaskExecutor;
+import org.apache.iotdb.db.pipe.task.execution.PipeProcessorSubtaskExecutor;
+import org.apache.iotdb.db.pipe.task.execution.PipeSubtaskExecutorManager;
 import org.apache.iotdb.db.pipe.task.stage.PipeTaskConnectorStage;
 import org.apache.iotdb.db.pipe.task.stage.PipeTaskExtractorStage;
 import org.apache.iotdb.db.pipe.task.stage.PipeTaskProcessorStage;
