+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/subscription/execution/executor/SubscriptionSubtaskExecutor.java
@@ -21,7 +21,7 @@ package org.apache.iotdb.db.subscription.execution.executor;
 
 import org.apache.iotdb.commons.concurrent.ThreadName;
 import org.apache.iotdb.commons.subscription.config.SubscriptionConfig;
-import org.apache.iotdb.db.pipe.execution.PipeConnectorSubtaskExecutor;
+import org.apache.iotdb.db.pipe.task.execution.PipeConnectorSubtaskExecutor;
 
 public class SubscriptionSubtaskExecutor extends PipeConnectorSubtaskExecutor {
