+++ iotdb-core/datanode/src/test/java/org/apache/iotdb/db/pipe/execution/PipeSubtaskExecutorTest.java
@@ -19,7 +19,7 @@
 
 package org.apache.iotdb.db.pipe.execution;
 
-import org.apache.iotdb.commons.pipe.execution.executor.PipeSubtaskExecutor;
+import org.apache.iotdb.commons.pipe.task.execution.executor.PipeSubtaskExecutor;
 import org.apache.iotdb.commons.pipe.task.subtask.PipeSubtask;
 
 import org.junit.After;
