+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/pipe/task/progress/PipeEventCommitter.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.commons.pipe.progress;
+package org.apache.iotdb.commons.pipe.task.progress;
 
 import org.apache.iotdb.commons.pipe.event.EnrichedEvent;
 import org.apache.iotdb.pipe.api.event.Event;
